FIX added checking into sys.modules before trying to import base_globals
during unpickling, in order to recreate the previous environnement an object lived in,
the module under the base_globals name is re-imported. However, it
is not always possible, especially if base_globals pointed to a
temporary or locally defined module. This PR introduces a fix that
checks if the module still exists during unpickling