TST: Refactor subprocess_pickle_echo for string inspection
Separate out the retrieval of the pickle string from the subprocess with
the unpickling, so the string can be directly compared.

This patch also adds an add_env parameter to allow subprocesses to be
passed environment variables. The main use case under consideration is
setting PYTHONHASHSEED, allowing these functions to be used to confirm
(in)consistency of strings under different initial conditions.