Update whichmodule tests in cloudpickle to work with numpy 2.2.
1) test_non_module_object_passing_whichmodule_test was not testing what it was supposed to test,
because after some refactoring passing "name" to "_whichmodule" became mandatory for the module lookup to occur.
Changed the test to pass "func" name.

2) test_importing_multiprocessing_does_not_impact_whichmodule was no longer testing what it was supposed
to test after the numpy 2.2 release which caused the numpy symbols to have a module attached, and therefore
the module lookup was no longer occuring.
Changed the test to avoid dependency on numpy, instead testing with our own function.